requirements.txt
flask==3.0.3
line-bot-sdk==3.22.0
Pillow==10.3.0
pytesseract==0.3.13
requests==2.32.3
gunicorn==22.0.0
